requirements.txt
attrs==25.3.0
certifi==2025.1.31
h11==0.14.0
idna==3.10
MouseInfo==0.1.3
numpy==2.2.4
opencv-python==4.11.0.86
outcome==1.3.0.post0
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python3-xlib==0.15
pytweening==1.2.0
selenium==4.30.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
